+++ gatr/__init__.py
@@ -5,4 +5,4 @@ from .layers.mlp.config import MLPConfig
 from .nets.axial_gatr import AxialGATr
 from .nets.gatr import GATr
 
-__version__ = "1.2.0"
+__version__ = "1.2.2"