+++ gatr/primitives/dual.py
@@ -8,6 +8,7 @@ import torch
 
 from gatr.primitives.bilinear import outer_product
 from gatr.utils.einsum import cached_einsum
+from gatr.utils.misc import minimum_autocast_precision
 
 # Flag which reference join implementations we're using
 _USE_EFFICIENT_JOIN = True
@@ -246,6 +247,7 @@ def efficient_equivariant_join(
     return reference[..., [14]] * cached_einsum("i j k , ... j, ... k -> ... i", kernel, x, y)
 
 
+@minimum_autocast_precision(torch.float32)
 def join_norm(
     x: torch.Tensor, y: torch.Tensor, square=False, channel_sum=False, channel_weights=None
 ) -> torch.Tensor: