+++ tests/gatr/utils/test_compile_linear.py
@@ -1,3 +1,5 @@
+# Copyright (c) 2023 Qualcomm Technologies, Inc.
+# All rights reserved.
 from copy import deepcopy
 
 import pytest